+++ data.tf
@@ -3,38 +3,18 @@ data "aws_caller_identity" "audit" {
 }
 data "aws_caller_identity" "master" {}
 
-data "aws_iam_role" "monitor_iam_access_audit" {
-  for_each = toset([for identity in coalesce(var.monitor_iam_access, []) : identity.name if identity.type == "AssumedRole" && identity.account == "audit"])
+data "aws_cloudwatch_log_group" "cloudtrail_audit" {
   provider = aws.audit
-  name     = each.value
+  name     = "aws-controltower/CloudTrailLogs"
 }
 
-data "aws_iam_user" "monitor_iam_access_audit" {
-  for_each  = toset([for identity in coalesce(var.monitor_iam_access, []) : identity.name if identity.type == "IAMUser" && identity.account == "audit"])
-  provider  = aws.audit
-  user_name = each.value
-}
-
-data "aws_iam_role" "monitor_iam_access_logging" {
-  for_each = toset([for identity in coalesce(var.monitor_iam_access, []) : identity.name if identity.type == "AssumedRole" && identity.account == "logging"])
+data "aws_cloudwatch_log_group" "cloudtrail_logging" {
   provider = aws.logging
-  name     = each.value
+  name     = "aws-controltower/CloudTrailLogs"
 }
 
-data "aws_iam_user" "monitor_iam_access_logging" {
-  for_each  = toset([for identity in coalesce(var.monitor_iam_access, []) : identity.name if identity.type == "IAMUser" && identity.account == "logging"])
-  provider  = aws.logging
-  user_name = each.value
-}
-
-data "aws_iam_role" "monitor_iam_access_master" {
-  for_each = toset([for identity in coalesce(var.monitor_iam_access, []) : identity.name if identity.type == "AssumedRole" && identity.account == "master"])
-  name     = each.value
-}
-
-data "aws_iam_user" "monitor_iam_access_master" {
-  for_each  = toset([for identity in coalesce(var.monitor_iam_access, []) : identity.name if identity.type == "IAMUser" && identity.account == "master"])
-  user_name = each.value
+data "aws_cloudwatch_log_group" "cloudtrail_master" {
+  name = "aws-controltower/CloudTrailLogs"
 }
 
 data "aws_organizations_organization" "default" {}